refactor(config): parseEnvFile 을 env-file.ts 로 분리 — 600줄 CI 가드
main 머지로 env.ts 가 605줄 도달(Gate 3 초과). 독립 헬퍼 분리로 578줄. 동작 변경 없음.
